Add PDF reader functionality and indexing to Document model
This commit introduces the 'pdf-reader' gem to the project's dependencies. The addition is accompanied by changes in the Document model to enable indexing functionality. After a Document instance is created, an after_create_commit callback triggers the indexing of the file's content into the search engine. A new 'index_file!' method has been added for this purpose, which uses the new 'file_loader' method to read the file. The 'file_loader' uses the 'Langchain::Loader' to load the document content from a URL instead of a previously used file path. The URL is obtained using Rails' URL helper in the 'file_url' method. Server host default URL options have been set in the development environment configuration to facilitate proper URL generation for locally stored files.