Refactor code for indexing document file
The method `index_file` in `document.rb` now divides the file into chunks and the text content is added to Vector search client. In case of successful indexing, the `indexed` field is updated to `true`. The `indexed` field was also added to the documents table in the database schema. This was needed to keep track of which documents have been indexed already.